buck2_critical_path

Function compute_critical_path_potentials

pub fn compute_critical_path_potentials(
    deps: &Graph,
    weights: &VertexData<u64>,
) -> Result<(CriticalPathVertexData<VertexId>, PathCost, CriticalPathVertexData<PathCost>, CriticalPathAccessor)>

This returns:

The critical path, as a list of vertices.

Its cost.

For each of the vertices in the critical path, the new critical path cost if that were vertex’s runtime was 0.

An accessor to obtain critical path for arbitary vertices.